+++ example/lib/main.dart
@@ -44,8 +44,8 @@ class _HomeState extends State<Home> {
         },
       ),
       bottomNavigationBar: BottomBar(
+        textStyle: TextStyle(fontWeight: FontWeight.bold),
         selectedIndex: _currentPage,
-        showActiveBackgroundColor: true,
         onTap: (int index) {
           _pageController.jumpToPage(index);
           setState(() => _currentPage = index);
@@ -55,23 +55,25 @@ class _HomeState extends State<Home> {
             icon: Icon(Icons.home),
             title: Text('Home'),
             activeColor: Colors.blue,
+            activeTitleColor: Colors.blue.shade600,
           ),
           BottomBarItem(
             icon: Icon(Icons.favorite),
             title: Text('Favorites'),
             activeColor: Colors.red,
-            darkActiveColor: Colors.red.shade400,
           ),
           BottomBarItem(
             icon: Icon(Icons.person),
             title: Text('Account'),
+            backgroundColorOpacity: 0.1,
             activeColor: Colors.greenAccent.shade700,
-            darkActiveColor: Colors.greenAccent.shade400,
           ),
           BottomBarItem(
             icon: Icon(Icons.settings),
             title: Text('Settings'),
             activeColor: Colors.orange,
+            activeIconColor: Colors.orange.shade600,
+            activeTitleColor: Colors.orange.shade700,
           ),
         ],
       ),